requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
distro==1.9.0
    # via openai
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.10.0
    # via openai
openai==1.97.0
    # via -r requirements.in
pydantic==2.11.7
    # via openai
pydantic-core==2.33.2
    # via pydantic
pydub==0.25.1
    # via -r requirements.in
python-dotenv==1.1.1
    # via -r requirements.in
requests==2.32.4
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   anyio
    #   openai
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   openai
typing-extensions==4.14.1
    # via
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via requests
